@@ -1,33 +1,51 @@
 {
-  "name": "PS212",
-  "version": "0.1.0",
+  "name": "ps212",
+  "version": "2.0.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint",
     "fix": "prettier --write \"src/**/*.{js,vue}\"",
-    "deploy": "NODE_ENV=production npm run build; push-dir --dir=dist --branch=gh-pages --cleanup"
+    "deploy": "NODE_ENV=production npm run build; push-dir --dir=dist --branch=gh-pages --cleanup",
+    "dev": "concurrently \"npm run dev:serve --silent\" \"npm run dev:build --silent\" \"npm run dev:lint --silent\" \"npm run dev:fix --silent\" \"npm run dev:deploy --silent\"",
+    "dev:serve": "vue-cli-service serve",
+    "dev:build": "vue-cli-service build",
+    "dev:lint": "vue-cli-service lint",
+    "dev:fix": "prettier --write \"src/**/*.{js,vue}\"",
+    "dev:deploy": "NODE_ENV=development npm run build"
   },
   "dependencies": {
-    "@directus/sdk-js": "^4.0.0-rc.1",
-    "moment": "^2.22.2",
-    "vue": "^2.5.16",
-    "vue-router": "^3.0.1",
-    "vue-typer": "^1.2.0",
-    "vue2-smooth-scroll": "^1.0.3",
-    "vuex": "^3.0.1"
+    "axios": "^1.3.0",
+    "dotenv": "^16.0.3",
+    "moment": "^2.29.4",
+    "vue": "^3.2.45",
+    "vue-router": "^4.1.6",
+    "vue3-typer": "^1.0.0",
+    "vue3-smooth-scroll": "^0.8.1",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.0.0-rc.9",
-    "@vue/cli-plugin-eslint": "^3.0.0-rc.9",
-    "@vue/cli-service": "^3.0.0-rc.9",
-    "node-sass": "^4.9.2",
-    "prettier": "^1.15.3",
+    "@babel/eslint-parser": "^7.19.1",
+    "@npmcli/fs": "^3.1.0",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-plugin-eslint": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "concurrently": "^7.6.0",
+    "css-loader": "^6.7.3",
+    "eslint": "^8.33.0",
+    "eslint-plugin-vue": "^9.9.0",
+    "node-polyfill-webpack-plugin": "^2.0.1",
+    "sass": "^1.58.3",
+    "prettier": "^2.8.3",
     "push-dir": "^0.4.1",
-    "sass-loader": "^7.0.3",
-    "vue-template-compiler": "^2.5.16",
-    "webpack": "^4.16.3"
+    "sass-loader": "^13.2.0",
+    "style-loader": "^3.3.1",
+    "vue-loader": "^17.0.1",
+    "vue-template-compiler": "^2.7.14",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
   },
   "eslintConfig": {
     "root": true,
@@ -40,7 +58,7 @@
     ],
     "rules": {},
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser"
     }
   },
   "postcss": {
