@@ -2,10 +2,11 @@
   <div id="app">
     <v-header msg="Pass this value" />
 
-    <transition name="fade">
-      <router-view></router-view>
-      <!-- component matched by the route will render here -->
-    </transition>
+    <router-view v-slot="{ Component }">
+      <transition>
+        <component :is="Component" />
+      </transition>
+    </router-view>
 
     <v-footer msg="Pass this value" />
   </div>
@@ -14,24 +15,33 @@
 <script>
 import VHeader from "./components/header.vue";
 import VFooter from "./components/footer.vue";
+import { onMounted, onUnmounted } from "vue";
 
 export default {
   name: "app",
   components: {
     VHeader,
-    VFooter
+    VFooter,
   },
-  created: function() {
-    window.addEventListener("scroll", function() {
-      var scroll = this.scrollY;
-      var element = document.getElementById("header");
+  setup() {
+    onMounted(() => {
+      window.addEventListener("scroll", handleScroll);
+    });
+
+    onUnmounted(() => {
+      window.removeEventListener("scroll", handleScroll);
+    });
+
+    const handleScroll = () => {
+      let scroll = window.scrollY;
+      let element = document.getElementById("header");
       if (scroll > 20) {
         element.classList.add("shadow");
       } else {
         element.classList.remove("shadow");
       }
-    });
-  }
+    };
+  },
 };
 </script>
 
