@@ -12,6 +12,8 @@
 </template>
 
 <script>
+import { computed } from "vue";
+
 export default {
   name: "v-career",
   props: {
@@ -19,31 +21,34 @@ export default {
     location: String,
     title: String,
     text: String,
-    color: String
+    color: String,
   },
-  computed: {
-    subject: function() {
-      return encodeURIComponent(this.title);
-    }
-  }
+  setup(props) {
+    const encodedTitle = computed(() => {
+      return encodeURIComponent(props.title);
+    });
+
+    return {
+      encodedTitle,
+    };
+  },
 };
 </script>
 
 <style lang="scss" scoped>
-@import "../assets/_variables.scss";
+@import "~@/assets/_variables.scss";
+
 .career {
-  .container {
-    // padding-bottom: 80px;
-    // border-bottom: 1px solid $light-gray;
-  }
   h2 {
     max-width: 800px;
     margin: 20px 0 10px;
   }
+
   div.body {
     max-width: 700px;
     margin-top: 0px;
   }
+
   a.button {
     background-color: $accent;
     display: inline-block;
@@ -55,7 +60,8 @@ export default {
 </style>
 
 <style lang="scss">
-@import "../assets/_variables.scss";
+@import "~@/assets/_variables.scss";
+
 .career {
   div.body {
     p {
@@ -63,10 +69,12 @@ export default {
     }
   }
 }
+
 .career ul {
   margin: 0 0 10px 0;
   padding: 0;
   list-style-type: disc;
+
   li {
     margin-top: 10px;
     font-family: "Gotham HTF", mono;
@@ -75,6 +83,7 @@ export default {
     color: $gray;
   }
 }
+
 @media only screen and (max-width: 800px) {
   .career ul {
     margin: 0 0 10px 10px;
